+++ osmaxx/conversion/__init__.py
@@ -1,1 +1,5 @@
+from .constants import output_format
+
 default_app_config = 'osmaxx.conversion.apps.ConversionConfig'
+
+__all__ = ['default_app_config', 'output_format']
