+++ osmaxx/conversion/models.py
@@ -7,8 +7,8 @@ from django.utils.text import slugify
 from django.utils.translation import gettext_lazy as _
 from rest_framework.reverse import reverse
 
+from osmaxx.conversion import output_format
 from osmaxx.clipping_area.models import ClippingArea
-from osmaxx.conversion.constants import output_format
 from osmaxx.conversion.converters.converter import convert
 from osmaxx.conversion.converters.converter_gis.detail_levels import DETAIL_LEVEL_CHOICES, DETAIL_LEVEL_ALL
 from osmaxx.conversion.constants.coordinate_reference_systems import CRS_CHOICES
