+++ osmaxx/conversion/serializers.py
@@ -1,8 +1,8 @@
 from rest_framework import serializers
 
+from osmaxx.conversion import output_format
 from osmaxx.conversion.converters.converter_gis import detail_levels
 from osmaxx.conversion.size_estimator import size_estimation_for_format
-from osmaxx.conversion.constants import output_format
 from .models import Job, Parametrization
