+++ osmaxx/excerptexport/forms/order_options_mixin.py
@@ -2,8 +2,9 @@ from crispy_forms.layout import Fieldset, Div
 from django import forms
 from django.utils.translation import ugettext_lazy as _
 
+from osmaxx.conversion import output_format
 from osmaxx.conversion.converters.converter_gis.detail_levels import DETAIL_LEVEL_CHOICES
-from osmaxx.conversion.constants import output_format, coordinate_reference_systems as crs
+from osmaxx.conversion.constants import coordinate_reference_systems as crs
 
 
 class OrderOptionsMixin(forms.Form):
