+++ tests/excerptexport/test_conversion_api_client.py
@@ -8,8 +8,8 @@ from hamcrest import assert_that, contains_inanyorder
 from requests import HTTPError
 from rest_framework.reverse import reverse
 
+from osmaxx.conversion import output_format
 from osmaxx.api_client import ConversionApiClient, API_client
-from osmaxx.conversion.constants import output_format
 from osmaxx.conversion.constants.statuses import RECEIVED
 from osmaxx.excerptexport.models import Excerpt, ExtractionOrder
 from osmaxx.job_progress.views import tracker
