+++ tests/excerptexport/test_models.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import User
 from django.contrib.gis import geos
 from hamcrest import assert_that, contains_inanyorder as contains_in_any_order
 
-from osmaxx.conversion.constants import output_format
+from osmaxx.conversion import output_format
 from osmaxx.conversion.constants.statuses import FINISHED
 from osmaxx.excerptexport import models
